[fix] fix the type of epoch_for_2stage

+++ train.py
@@ -27,7 +27,7 @@ def parse_params():
     parser.add_argument('--warmup_epochs', default=5, type=int)
     parser.add_argument('--without_schedule', action='store_false')
     parser.add_argument('--lr_for_stage2', default=0.000008, type=float)
-    parser.add_argument('--epoch_for_2stage', default=25, type=float)
+    parser.add_argument('--epoch_for_2stage', default=25, type=int)
     parser.add_argument('--detail_summary', action='store_true')
     parser.add_argument('--imbalance', action='store_true')
